Prototype migration of Exult to SDL 3 : Followup fixes for iOS
  Commit 1 of 1 :
    exult.cc
      Align SDL_GetDesktopDisplayMode call on imagewin/imagewin.cc
    imagewin/save_screenshot.cc
      Alter SDL_RWwrite calls to the new signature, one argument less
    ios/GamePadView.m + android/lib/src/TouchUI_Android.cc
      SDL 2 device index becomes SDL 3 device ID
    audio/Audio.cc + AudioChannel.cc + Midi.cc
      Fix timestamp format in debug messages

+++ audio/Audio.cc
@@ -79,7 +79,10 @@ const int* Audio::bg2si_songs = nullptr;
 inline char* formatTicks() {
 	static char formattedTicks[32];
 	uint64      ticks = SDL_GetTicks();
-	snprintf(formattedTicks, 32, "[ %5ld.%0ld ] ", ticks / 1000, ticks % 1000);
+	snprintf(
+			formattedTicks, 32, "[ %5u.%03u ] ",
+			static_cast<uint32>(ticks / 1000),
+			static_cast<uint32>(ticks % 1000));
 	return formattedTicks;
 }
 #endif
