+++ audio/Midi.cc
@@ -54,7 +54,10 @@ using std::string;
 inline char* formatTicks() {
 	static char formattedTicks[32];
 	uint64      ticks = SDL_GetTicks();
-	snprintf(formattedTicks, 32, "[ %5ld.%0ld ] ", ticks / 1000, ticks % 1000);
+	snprintf(
+			formattedTicks, 32, "[ %5u.%03u ] ",
+			static_cast<uint32>(ticks / 1000),
+			static_cast<uint32>(ticks % 1000));
 	return formattedTicks;
 }
 #endif
